Define journal settings in journal crate
