+++ crates/journal/Cargo.toml
@@ -13,9 +13,12 @@ editor = { path = "../editor" }
 gpui = { path = "../gpui" }
 util = { path = "../util" }
 workspace = { path = "../workspace" }
+settings = { path = "../settings" }
+
 anyhow.workspace = true
 chrono = "0.4"
 dirs = "4.0"
+serde.workspace = true
+schemars.workspace = true
 log.workspace = true
-settings = { path = "../settings" }
 shellexpand = "2.1.0"
